requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pdfplumber>=0.9.0
pytest>=7.4.0
openai>=1.3.0
google-generativeai>=0.3.0
groq>=0.4.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0
requests>=2.31.0
